Create FeatureSelectionMixin for shared [inverse_]transform code
